Initial refactoring - Clean Architecture
