+++ app/config/__init__.py
@@ -1,0 +1,5 @@
+from app.config.data_module import DataModule
+from app.config.use_case_module import UseCaseModule
+from app.config.domain_module import DomainModule
+
+__all__ = ['DataModule', 'UseCaseModule', 'DomainModule']
