--- app/controller/__init__.py
@@ -1,4 +1,0 @@
-from app.controller.base_controller import BaseController
-from app.controller.cluster_group import ClusterGroupController
-
-__all__ = ['BaseController', 'ClusterGroupController']
