--- app/data/datasource/__init__.py
@@ -1,5 +1,0 @@
-from app.data.datasource.datasource import ClusterDatasource
-from app.data.datasource.locator import ClusterDataSourceLocator
-
-__all__ = ['ClusterDatasource', 'ClusterDataSourceLocator']
-
