--- app/data/repository/__init__.py
@@ -1,4 +1,0 @@
-from app.data.repository.enforcement import EnforcementRepository
-from app.data.repository.cluster import ClusterRepository
-
-__all__ = ['EnforcementRepository', 'ClusterRepository']
