--- app/di/__init__.py
@@ -1,4 +1,0 @@
-from app.di.argo_module import ArgoModule
-from app.di.use_case_module import UseCaseModule
-
-__all__ = ['ArgoModule', 'UseCaseModule']
