+++ app/domain/source_locator.py
@@ -1,0 +1,7 @@
+from app.domain.repositories import SourceRepository
+from app.domain.entities import EnforcementSource
+
+
+class SourceLocator:
+    def locate(self, source: EnforcementSource) -> SourceRepository:
+        pass
