+++ app/domain/use_case/sync_rules.py
@@ -1,0 +1,8 @@
+import attr
+from typing import List
+
+
+@attr.s(auto_attribs=True)
+class SyncRulesUseCase:
+    pass
+
