--- app/helper/__init__.py
@@ -1,3 +1,0 @@
-from app.helper.logger import logger
-
-__all__ = ['Config', 'logger']
