+++ app/infra/__init__.py
@@ -1,0 +1,3 @@
+from app.infra.logger import logger
+
+__all__ = ['Config', 'logger']
