--- app/use_case/__init__.py
@@ -1,3 +1,0 @@
-from app.use_case.apply_rules import ApplyRulesUseCase
-
-__all__ = ['ApplyRulesUseCase']
